chore: app design restructure
